Split benchmarks
